construcao gulosa vizinho mais proximo

+++ src/MS.cpp
@@ -20,7 +20,7 @@ float MS(int n, std::vector<int> &s, float **d, int iter_max)
     clock_t inicio_CPU, fim_CPU;
     vector<int> s_star;
 
-    fo = fo_star = INT_MAX;
+    fo = fo_star = INT8_MAX;
     limpa_arquivo((char*)"MS.txt");
     inicio_CPU = fim_CPU = clock();
     imprime_fo((char*)"MS.txt", (fim_CPU - inicio_CPU)/CLOCKS_PER_SEC,fo_star,0);
